Bug 25119: Fix account type and description for individual writeoffs/payments
When paying or writing off a single line from the list of pending charges,
the account type displayed the code instead of the description.

And when bug 23483 fixed the display of the description, 2 variables where
missed. This change is only visible when looking at the page source:

<input type="hidden" name="description" id="description" value="" />

With the patch applied, the value will be filled with the description.

To test:
- Create one or more manual invoices of different types
- Try to pay/write off using the buttons within the table
- Verify the account type shows as code, not description
- Apply patch
- Repeat
- Bonus points: Check the source code for the hidden input with description.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>